# Conflicts:
#	first-spring-step/src/main/java/com/jamedow/laodoufang/service/impl/ElasticSearchServiceImpl.java
#	first-spring-step/src/main/resources/development/generatorConfig.xml

+++ first-spring-step/src/main/java/com/jamedow/laodoufang/web/UsersController.java
@@ -90,7 +90,7 @@ public class UsersController {
                 return UsersService.USER_EXIST;//用户已存在
 
             user = new Users();
-            user.setId((long) (Constant.ZERO));
+            user.setId(Constant.ZERO);
             user.setUsername(userName);
             password = MD5.md5crypt(password);
             user.setPassword(password);
